progress on melody algo and push_back, minor enhancements

+++ src/utils.c
@@ -6,14 +6,14 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 10:16:48 by estettle          #+#    #+#             */
-/*   Updated: 2024/12/11 12:05:16 by estettle         ###   ########.fr       */
+/*   Updated: 2024/12/11 12:47:03 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/push_swap.h"
 
 /**
- * @brief Converts a number as a string into its 64 bit integer representation.
+ * @brief Converts a number as a string into its 64-bit integer representation.
  * This function accepts any number of white space (isspace(3)) as a suffix,
  * as well as one minus (-) or plus (-) sign.
  *
@@ -33,7 +33,7 @@ int64_t	ft_atol(const char *str)
 		i++;
 	if (str[i] == '-' || str[i] == '+')
 		if (str[i++] == '-')
-			sign = -sign;
+			sign = (int8_t)-sign;
 	while (str[i] >= '0' && str[i] <= '9')
 	{
 		converted += str[i++] - '0';
